pyproject.toml
[tool.poetry]
name = "OpenAI_Agent_Swarm"
version = "0.0.0"
description = "test playground for OpenAI_Agent_Swarm fork/clone"
authors = ["Kevin <kevin.armengol@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1, <3.12"
openai = "^1.2.0"
llama-index = "^0.8.65"
pillow = "^10.1.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.26.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
